refactor(scripts): hoist utils/prettier/* up to prettier/*
Looking at some other tasks that I am going to have to do, such as:

https://github.com/liferay/liferay-npm-tools/issues/309

I want to move some things around. We already have a top-level "jest"
folder where we put "Jest" stuff; so, this commit moves our "Prettier"
stuff up into a top-level "prettier" folder.

In a separate PR I will rename the "format" directory to "jsp", because
that is currently the home of our JSP-formatting code, but it will soon
contain some JSP-linting stuff too.
