+++ packages/liferay-npm-scripts/src/prettier/rules/newline-before-block-statements.js
