+++ packages/liferay-npm-scripts/test/prettier/index.js
@@ -4,9 +4,9 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-const config = require('../../../src/config/prettier.json');
-const prettier = require('../../../src/utils/prettier');
-const dedent = require('../../../support/dedent');
+const config = require('../../src/config/prettier.json');
+const prettier = require('../../src/prettier');
+const dedent = require('../../support/dedent');
 
 /**
  * Helper function to make code samples in this file a little more readable:
@@ -68,7 +68,7 @@ describe('code``', () => {
 	});
 });
 
-describe('utils/prettier/index.js', () => {
+describe('prettier/index.js', () => {
 	describe('prettier.format()', () => {
 		function format(source, options = {}) {
 			return prettier.format(source, {
