+++ packages/liferay-npm-scripts/test/scripts/format.js
@@ -9,9 +9,9 @@ const os = require('os');
 const path = require('path');
 
 const formatJSP = require('../../src/jsp/formatJSP');
+const prettier = require('../../src/prettier');
 const format = require('../../src/scripts/format');
 const log = require('../../src/utils/log');
-const prettier = require('../../src/utils/prettier');
 
 jest.mock('../../src/jsp/formatJSP');
 jest.mock('../../src/utils/log');
